Refactor locking package structure.
Moved locking-related classes from `system.notification.email.locking` to `system.locking` for better organization and clarity. Updated all imports and references accordingly.
