+++ src/main/java/info/mackiewicz/bankapp/system/locking/LockingConfig.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.notification.email.locking;
+package info.mackiewicz.bankapp.system.locking;
 
 import org.springframework.boot.context.properties.ConfigurationProperties;
