+++ src/main/java/info/mackiewicz/bankapp/system/locking/LockingStrategy.java
@@ -1,4 +1,4 @@
-package info.mackiewicz.bankapp.system.notification.email.locking;
+package info.mackiewicz.bankapp.system.locking;
 
 import info.mackiewicz.bankapp.core.account.exception.AccountLockException;
